Figured out issue with the occlusion case.
The occluding surface needed to be included in the
RadiationHeatTransferSetup. Had to do some hacking to figure out
how to get a handle on the cylinder's surface. Now the solution
is close to the expected analytic solution.

It should be noted that the VF from the cylinder's surface to itself
is not zero for coarse meshes. This is because the coarse mesh is
slightly concave. This should go away with better meshes. Need
to study this to make sure.